+++ api/dependencies.py
@@ -1,4 +1,8 @@
+from collections.abc import AsyncGenerator
+from typing import Any
+
 from fastapi import Depends
+import redis.asyncio as redis
 from redis.asyncio import Redis
 from sqlalchemy.ext.asyncio import AsyncSession
 
@@ -14,20 +18,15 @@ from api.domain.provider import (
 from api.domain.role import LimitRepository, PermissionRepository
 from api.domain.router import RouterRateLimiter
 from api.domain.usage import UsageRecorder
-from api.domain.user import UserPasswordEncoder
+from api.domain.user import AuthenticatedUserQuery, UserPasswordEncoder
 from api.infrastructure.bcrypt import BcryptUserPasswordEncoder
-from api.infrastructure.dependencies import (
-    _key_encoder,
-    _key_repository,
-    get_postgres_session,
-    get_redis_client,
-)
 from api.infrastructure.ecologit import EcologitModelEnvironmentalImpactsComputer
-from api.infrastructure.fastapi import RequestContextUsageRecorder, request_context
+from api.infrastructure.fastapi import RequestContextUsageRecorder
+from api.infrastructure.fastapi.dependencies import request_context
 from api.infrastructure.http import HttpProviderAdapterBuilder, HttpProviderClient
-from api.infrastructure.model import ModelProviderGateway
 from api.infrastructure.jwt import JwtKeyEncoder
 from api.infrastructure.postgres import (
+    PostgresAuthenticatedUserQuery,
     PostgresKeyRepository,
     PostgresLimitRepository,
     PostgresPermissionRepository,
@@ -60,7 +59,36 @@ from api.utils.configuration import configuration
 from api.utils.context import global_context
 
 
+# databases
+async def get_postgres_session() -> AsyncGenerator[AsyncSession]:
+    session_factory = global_context.postgres_session_factory
+    async with session_factory() as postgres_session:
+        try:
+            yield postgres_session
+            if postgres_session.in_transaction():
+                await postgres_session.commit()
+        except Exception:
+            if postgres_session.in_transaction():
+                await postgres_session.rollback()
+            raise
+
+
+async def get_redis_client() -> AsyncGenerator[Redis, Any]:
+    client = redis.Redis(connection_pool=global_context.redis_pool)
+    yield client
+    await client.aclose()
+
+
+# queries
+def _authenticated_user_query(session: AsyncSession = Depends(get_postgres_session)) -> AuthenticatedUserQuery:
+    return PostgresAuthenticatedUserQuery(postgres_session=session)
+
+
 # helpers
+def _key_encoder() -> KeyEncoder:
+    return JwtKeyEncoder(secret_key=configuration.settings.auth_secret_key)
+
+
 def _model_tokenizer() -> ModelTokenizer:
     return TiktokenModelTokenizer(model=global_context._tokenizer)
 
@@ -127,6 +155,10 @@ def _provider_capabilities_probe(
     provider_adapter_builder: ProviderAdapterBuilder = Depends(_provider_adapter_builder),
 ) -> ProviderCapabilitiesProbe:
     return ProviderCapabilitiesProbe(provider_client=provider_client, provider_adapter_builder=provider_adapter_builder)
+
+
+def _key_repository(key_encoder: KeyEncoder = Depends(_key_encoder), session: AsyncSession = Depends(get_postgres_session)) -> KeyRepository:
+    return PostgresKeyRepository(key_encoder=key_encoder, postgres_session=session)
 
 
 # health use cases
