+++ api/infrastructure/fastapi/_requestcontextusagerecorder.py
@@ -1,7 +1,8 @@
 from contextvars import ContextVar
 
 from api.domain.usage import UsageRecorder
-from api.infrastructure.fastapi._requestcontext import RequestContext
+
+from ._requestcontext import RequestContext
 
 
 class RequestContextUsageRecorder(UsageRecorder):
