+++ api/infrastructure/fastapi/accesscontroller.py
@@ -6,12 +6,13 @@ from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from jose import JWTError
 from pydantic import ValidationError
 
+from api.dependencies import _authenticated_user_query, _key_repository
 from api.domain.key import KeyEncoder, KeyRepository
 from api.domain.key.entities import Key
 from api.domain.key.errors import KeyNotFoundError
 from api.domain.user import AuthenticatedUserQuery
-from api.infrastructure.dependencies import _authenticated_user_query, _key_repository
-from api.infrastructure.fastapi._requestcontext import RequestContext, get_request_context
+from api.infrastructure.fastapi._requestcontext import RequestContext
+from api.infrastructure.fastapi.dependencies import get_request_context
 from api.infrastructure.fastapi.endpoints.exceptions import (
     AccountExpiredHTTPException,
     InvalidAPIKeyHTTPException,
