+++ api/tests/integration/fastapi/test_accesscontroller.py
@@ -7,8 +7,9 @@ from jose import jwt
 import pytest
 
 from api.domain.role.entities import PermissionType
-from api.infrastructure.fastapi import RequestContext, request_context
-from api.infrastructure.fastapi._accesscontroler import AccessController
+from api.infrastructure.fastapi import RequestContext
+from api.infrastructure.fastapi.accesscontroller import AccessController
+from api.infrastructure.fastapi.dependencies import request_context
 from api.infrastructure.fastapi.endpoints.exceptions import (
     AccountExpiredHTTPException,
     InvalidAPIKeyHTTPException,
